+++ src/main/kotlin/dev/thiagosouto/plugins/bom/models.kt
@@ -1,8 +1,9 @@
 package dev.thiagosouto.plugins.bom
 
 internal data class BomInfo(val projectAttributes: List<Tag>, val tags: List<Tag>, val dependencies: List<Dependency>)
-internal class SimpleTag(name: String, val value: String) : Tag(name)
-internal class RootTag(name: String, val tags: List<Tag>) : Tag(name)
+internal class SimpleTag(override val name: String, val value: String) : Tag
+
+internal class RootTag(override val name: String, val tags: List<Tag>) : Tag
 
 internal data class Dependency(
     val artifactId: String,
@@ -12,4 +13,6 @@ internal data class Dependency(
 )
 
 internal data class Exclusion(val artifactId: String, val groupId: String)
-internal abstract class Tag(val name: String)
+interface Tag {
+    val name: String
+}