+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 ### Added
+- Added developerId, developerName, developerEmail, scmConnection, scmDeveloperConnection, scmUrl fields to `BomMetadata`.
+- Added `licenses` tag to generated `pom.xml` when the user fills licenses info in `BomMetadata`.
+
 ### Modified
 ### Removed